Don't modify the input alignment file.
Previously we modified the alignment file to remove BLK
residues if the target contained them but the first template
did not. It's not clear why this was done, but in cases where
this happens, the generated models end up with fewer atoms than
those made with the full alignment, so that the atom numbering
in the restraints file is incorrect.